[edit] server tests refactoring
